+++ src/org/schema/enumerations/events/EventMovedOnline.php
@@ -2,13 +2,14 @@
 
 namespace org\schema\enumerations\events;
 
-use org\schema\Enumeration;
-
 /**
  * Indicates that the event was changed to allow online participation.
+ *
+ * See eventAttendanceMode for specifics of whether it is now fully or partially online.
+ *
  * @see https://schema.org/EventMovedOnline
  */
-class EventMovedOnline extends Enumeration
+class EventMovedOnline extends EventStatusType
 {
 
 }