+++ src/org/schema/enumerations/events/EventPostponed.php
@@ -2,14 +2,14 @@
 
 namespace org\schema\enumerations\events;
 
-use org\schema\Enumeration;
-
 /**
  * The event has been postponed and no new date has been set.
+ *
  * The event's previousStartDate should be set.
+ *
  * @see https://schema.org/EventPostponed
  */
-class EventPostponed extends Enumeration
+class EventPostponed extends EventStatusType
 {
 
 }