+++ src/org/schema/enumerations/events/EventScheduled.php
@@ -2,13 +2,14 @@
 
 namespace org\schema\enumerations\events;
 
-use org\schema\Enumeration;
-
 /**
- * The event is taking place or has taken place on the startDate as scheduled. Use of this value is optional, as it is assumed by default.
+ * The event is taking place or has taken place on the startDate as scheduled.
+ *
+ * Use of this value is optional, as it is assumed by default.
+ *
  * @see https://schema.org/EventScheduled
  */
-class EventScheduled extends Enumeration
+class EventScheduled extends EventStatusType
 {
 
 }